# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AutoSSL 证书管理API文档",
        "title": "AutoSSL API",
        "termsOfService": "http://example.com/terms",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "http://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/apiv1",
    "paths": {
...
        "/admin/users": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "管理员获取系统所有用户列表（需要管理员权限）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员"
                ],
                "summary": "获取所有用户",
                "responses": {
                    "200": {
                        "description": "成功响应",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/ent.User"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/users/role": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "管理员修改指定用户的角色（需要管理员权限）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员"
                ],
                "summary": "更新用户角色",
                "responses": {
                    "200": {
                        "description": "成功响应",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "无效请求格式",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "access.Type": {
            "type": "string",
            "enum": [
                "ssh",
                "api",
                "tencentcloud-teo",
                "other"
            ],
            "x-enum-varnames": [
                "TypeSSH",
                "TypeAPI",
                "TypeTencentcloudTeo",
                "TypeOther"
            ]
        },
...
        "certificate.ValidationType": {
            "type": "string",
            "enum": [
                "dns",
                "http"
            ],
            "x-enum-varnames": [
                "ValidationTypeDNS",
                "ValidationTypeHTTP"
            ]
        },
...
        "ent.ACMEAccount": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the ACMEAccountQuery when eager-loading is set.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.ACMEAccountEdges"
                        }
                    ]
                },
                "email": {
                    "description": "ACME账号邮箱地址",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "private_key": {
                    "description": "ACME账号私钥，PEM格式",
                    "type": "string"
                },
                "provider": {
                    "description": "SSL证书提供商: letsencrypt, zerossl",
                    "type": "string"
                },
                "registration_body": {
                    "description": "ACME注册响应体，JSON格式",
                    "type": "string"
                },
                "registration_uri": {
                    "description": "ACME注册URI",
                    "type": "string"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
        "ent.ACMEAccountEdges": {
            "type": "object",
            "properties": {
                "certificates": {
                    "description": "Certificates holds the value of the certificates edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Certificate"
                    }
                },
                "user": {
                    "description": "User holds the value of the user edge.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.User"
                        }
                    ]
                }
            }
        },
        "ent.Access": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "CreatedAt holds the value of the \"created_at\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the AccessQuery when eager-loading is set.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.AccessEdges"
                        }
                    ]
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "is_active": {
                    "description": "是否启用",
                    "type": "boolean"
                },
                "name": {
                    "description": "权限配置名称",
                    "type": "string"
                },
                "type": {
                    "description": "访问类型",
                    "allOf": [
                        {
                            "$ref": "#/definitions/access.Type"
                        }
                    ]
                },
                "updated_at": {
                    "description": "UpdatedAt holds the value of the \"updated_at\" field.",
                    "type": "string"
                }
            }
        },
        "ent.AccessEdges": {
            "type": "object",
            "properties": {
                "deployments": {
                    "description": "部署记录",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.CertificateDeployment"
                    }
                },
                "server_group_members": {
                    "description": "关联的服务器组成员",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.ServerGroupMember"
                    }
                },
                "user": {
                    "description": "所属用户",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.User"
                        }
                    ]
                }
            }
        },
        "ent.Certificate": {
            "type": "object",
            "properties": {
                "access_ids": {
                    "description": "访问配置ID列表",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "acme_account_email": {
                    "description": "ACME账号邮箱",
                    "type": "string"
                },
                "created_at": {
                    "description": "CreatedAt holds the value of the \"created_at\" field.",
                    "type": "string"
                },
                "dns_provider_name": {
                    "description": "DNS提供商名称",
                    "type": "string"
                },
                "domains": {
                    "description": "证书域名列表",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the CertificateQuery when eager-loading is set.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.CertificateEdges"
                        }
                    ]
                },
                "error_message": {
                    "description": "错误信息",
                    "type": "string"
                },
                "expired_at": {
                    "description": "过期时间",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "issued_at": {
                    "description": "签发时间",
                    "type": "string"
                },
                "issued_by": {
                    "description": "签发机构",
                    "type": "string"
                },
                "renewal_days_before": {
                    "description": "距离过期前多少天开始续期",
                    "type": "integer"
                },
                "renewal_enabled": {
                    "description": "是否启用自动续期",
                    "type": "boolean"
                },
                "server_group_id": {
                    "description": "服务器组ID",
                    "type": "integer"
                },
                "status": {
                    "description": "Status holds the value of the \"status\" field.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/certificate.Status"
                        }
                    ]
                },
                "updated_at": {
                    "description": "UpdatedAt holds the value of the \"updated_at\" field.",
                    "type": "string"
                },
                "validation_type": {
                    "description": "验证类型",
                    "allOf": [
                        {
                            "$ref": "#/definitions/certificate.ValidationType"
                        }
                    ]
                }
            }
        },
...
        "ent.DNSProvider": {
            "type": "object",
            "properties": {
                "config": {
                    "description": "供应商配置 (JSON格式，包含所有认证信息)",
                    "type": "object",
                    "additionalProperties": true
                },
                "created_at": {
                    "description": "CreatedAt holds the value of the \"created_at\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the DNSProviderQuery when eager-loading is set.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.DNSProviderEdges"
                        }
                    ]
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "provider_type": {
                    "description": "DNS服务商类型",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dnsprovider.ProviderType"
                        }
                    ]
                },
                "updated_at": {
                    "description": "UpdatedAt holds the value of the \"updated_at\" field.",
                    "type": "string"
                }
            }
        },
...
        "ent.NotificationConfig": {
            "type": "object",
            "properties": {
                "config": {
                    "description": "通知配置参数(JSON格式)\n\t\t\t\t- webhook: {url:string, secret:string}\n\t\t\t\t- email: {smtp_server:string, smtp_port:int, username:string, password:string, from:string}\n\t\t\t\t- sms: {api_key:string, template_id:string, sign_name:string}",
                    "type": "object",
                    "additionalProperties": true
                },
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the NotificationConfigQuery when eager-loading is set.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.NotificationConfigEdges"
                        }
                    ]
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "is_active": {
                    "description": "是否启用",
                    "type": "boolean"
                },
                "name": {
                    "description": "配置名称",
                    "type": "string"
                },
                "type": {
                    "description": "通知类型",
                    "allOf": [
                        {
                            "$ref": "#/definitions/notificationconfig.Type"
                        }
                    ]
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
...
        "ent.Order": {
            "type": "object",
            "properties": {
                "cert_num": {
                    "description": "证书数量",
                    "type": "integer"
                },
                "created_at": {
                    "description": "CreatedAt holds the value of the \"created_at\" field.",
                    "type": "string"
                },
                "deploy_num": {
                    "description": "部署服务器数量",
                    "type": "integer"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the OrderQuery when eager-loading is set.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.OrderEdges"
                        }
                    ]
                },
                "end_date": {
                    "description": "订单失效时间",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "order_type": {
                    "description": "订单类型",
                    "allOf": [
                        {
                            "$ref": "#/definitions/order.OrderType"
                        }
                    ]
                },
                "price": {
                    "description": "订单金额",
                    "type": "number"
                },
                "start_date": {
                    "description": "订单生效时间",
                    "type": "string"
                },
                "status": {
                    "description": "Status holds the value of the \"status\" field.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/order.Status"
                        }
                    ]
                }
            }
        },
...
        "ent.Role": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "角色描述",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the RoleQuery when eager-loading is set.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.RoleEdges"
                        }
                    ]
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "is_system": {
                    "description": "是否为系统角色",
                    "type": "boolean"
                },
                "name": {
                    "description": "角色名称",
                    "type": "string"
                }
            }
        },
...
        "ent.ServerGroup": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "CreatedAt holds the value of the \"created_at\" field.",
                    "type": "string"
                },
                "description": {
                    "description": "组描述",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the ServerGroupQuery when eager-loading is set.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.ServerGroupEdges"
                        }
                    ]
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "组名称",
                    "type": "string"
                },
                "status": {
                    "description": "组状态",
                    "allOf": [
                        {
                            "$ref": "#/definitions/servergroup.Status"
                        }
                    ]
                },
                "type": {
                    "description": "组访问类型",
                    "allOf": [
                        {
                            "$ref": "#/definitions/servergroup.Type"
                        }
                    ]
                },
                "updated_at": {
                    "description": "UpdatedAt holds the value of the \"updated_at\" field.",
                    "type": "string"
                }
            }
        },
...
        "ent.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "CreatedAt holds the value of the \"created_at\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the UserQuery when eager-loading is set.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.UserEdges"
                        }
                    ]
                },
                "email": {
                    "description": "登录邮箱",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "is_active": {
                    "description": "账户状态",
                    "type": "boolean"
                },
                "is_admin": {
                    "description": "管理员标识",
                    "type": "boolean"
                },
                "updated_at": {
                    "description": "UpdatedAt holds the value of the \"updated_at\" field.",
                    "type": "string"
                }
            }
        },
        "ent.UserEdges": {
            "type": "object",
            "properties": {
                "access_configs": {
                    "description": "访问配置",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Access"
                    }
                },
                "acme_accounts": {
                    "description": "用户的ACME账户",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.ACMEAccount"
                    }
                },
                "certificates": {
                    "description": "用户证书",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Certificate"
                    }
                },
                "dns_providers": {
                    "description": "DNS提供商配置",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.DNSProvider"
                    }
                },
                "notification_configs": {
                    "description": "通知配置",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.NotificationConfig"
                    }
                },
                "orders": {
                    "description": "用户订单",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Order"
                    }
                },
                "roles": {
                    "description": "用户角色",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Role"
                    }
                },
                "server_groups": {
                    "description": "服务器组",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.ServerGroup"
                    }
                },
                "workflow_steps": {
                    "description": "工作流步骤",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.WorkflowStep"
                    }
                },
                "workflows": {
                    "description": "用户工作流",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Workflow"
                    }
                }
            }
        },
        "ent.Workflow": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "description": "完成时间",
                    "type": "string"
                },
                "created_at": {
                    "description": "CreatedAt holds the value of the \"created_at\" field.",
                    "type": "string"
                },
                "current_step": {
                    "description": "当前工作流步骤",
                    "type": "integer"
                },
                "domain": {
                    "description": "域名",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the WorkflowQuery when eager-loading is set.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.WorkflowEdges"
                        }
                    ]
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "status": {
                    "description": "工作流状态: pending, processing, issued, deploying, completed, failed, cancelled",
                    "type": "string"
                },
                "total_steps": {
                    "description": "总工作流步骤数",
                    "type": "integer"
                },
                "updated_at": {
                    "description": "UpdatedAt holds the value of the \"updated_at\" field.",
                    "type": "string"
                }
            }
        },
...
        "ent.WorkflowStep": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "description": "完成时间",
                    "type": "string"
                },
                "created_at": {
                    "description": "CreatedAt holds the value of the \"created_at\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the WorkflowStepQuery when eager-loading is set.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.WorkflowStepEdges"
                        }
                    ]
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "log": {
                    "description": "步骤日志",
                    "type": "string"
                },
                "name": {
                    "description": "步骤名称",
                    "type": "string"
                },
                "started_at": {
                    "description": "开始时间",
                    "type": "string"
                },
                "status": {
                    "description": "步骤状态",
                    "allOf": [
                        {
                            "$ref": "#/definitions/workflowstep.Status"
                        }
                    ]
                },
                "step_number": {
                    "description": "步骤编号",
                    "type": "integer"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}